flags location changed
