+++ cmd/server/flags.go
@@ -1,19 +1,1 @@
 package main
-
-import (
-	"flag"
-	"fmt"
-	"log"
-)
-
-var runServerAddrFlag string
-
-func parseServerFlags() {
-	flag.StringVar(&runServerAddrFlag, "a", ":8080", "server listens on this port")
-	flag.Parse()
-
-	if len(flag.Args()) > 0 {
-		fmt.Printf("Unknown flags %v\n", flag.Args())
-		log.Fatal("Error: unknown flags were given")
-	}
-}
